int -> isize in tests / benches
Fixes #67.

+++ benches/tokenizer.rs
@@ -7,7 +7,7 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-#![feature(box_syntax, core, env, int_uint, io, path, std_misc, start, test)]
+#![feature(box_syntax, core, env, io, path, std_misc, start, test)]
 
 extern crate test;
 extern crate html5ever;
@@ -148,7 +148,7 @@ fn tests() -> Vec<TestDescAndFn> {
 }
 
 #[start]
-fn start(argc: int, argv: *const *const u8) -> int {
+fn start(argc: isize, argv: *const *const u8) -> isize {
     unsafe {
         rt::args::init(argc, argv);
     }
